+++ RiskOfTheAncients2/Items/SparkOfCourage.cs
@@ -40,9 +40,9 @@ namespace ROTA2.Items
             ModSettingsManager.AddOption(new FloatFieldOption(DamageBase));
             DamagePerStack = configuration.Bind("Item: " + ItemName, "Stacking Damage Bonus", 15f, "How much damage should be provided by subsequent stacks?");
             ModSettingsManager.AddOption(new FloatFieldOption(DamagePerStack));
-            ArmorBase = configuration.Bind("Item: " + ItemName, "Initial Armor Bonus", 10f, "How much armor should be provided by the first stack?");
+            ArmorBase = configuration.Bind("Item: " + ItemName, "Initial Armor Bonus", 15f, "How much armor should be provided by the first stack?");
             ModSettingsManager.AddOption(new FloatFieldOption(ArmorBase));
-            ArmorPerStack = configuration.Bind("Item: " + ItemName, "Stacking Armor Bonus", 10f, "How much armor should be provided by subsequent stacks?");
+            ArmorPerStack = configuration.Bind("Item: " + ItemName, "Stacking Armor Bonus", 15f, "How much armor should be provided by subsequent stacks?");
             ModSettingsManager.AddOption(new FloatFieldOption(ArmorPerStack));
             HealthThreshold = configuration.Bind("Item: " + ItemName, "Health Threshold", 50f, "At what percent of maximum health should the bonus flip from damage to armor?");
             ModSettingsManager.AddOption(new FloatFieldOption(HealthThreshold));